+++ bombshell/__init__.py
@@ -1,3 +1,4 @@
 from .core import CommandChain, CompletedProcess, Pipeline, PipelineError, Process
+from .resources import ResourceData
 
-__all__ = ["Process", "Pipeline", "CompletedProcess", "PipelineError", "CommandChain"]
+__all__ = ["Process", "Pipeline", "CompletedProcess", "PipelineError", "CommandChain", "ResourceData"]